Read from the pipe in a separate thread
We start a separate thread in which we read from the
pipe. The data read on the pipe is forwarded to the main
thread from where we send the data to the client socket.

This will allow us to implement a ping on the main thread
that checks for the availability of the client connection.